[ContentBundle] improve functional tests
