+++ app/config/config.yml
@@ -71,7 +71,7 @@ doctrine:
         entity_managers:
             default:
                 naming_strategy: doctrine.orm.naming_strategy.underscore
-                auto_mapping: true
+                auto_mapping: false
                 metadata_cache_driver:
                     type: service
                     id: doctrine_cache.providers.main_cache
