--- src/SWP/Bundle/ContentBundle/Resources/config/serializer/Doctrine.ORM.Article.yml
@@ -1,2 +1,0 @@
-SWP\Bundle\ContentBundle\Doctrine\ORM\Article:
-    exclusion_policy: ALL
