+++ src/SWP/Bundle/ContentBundle/Resources/config/validation.yml
@@ -1,4 +1,4 @@
-SWP\Bundle\ContentBundle\Doctrine\ORM\Route:
+SWP\Bundle\ContentBundle\Model\Route:
     properties:
         type:
             - SWP\Bundle\ContentBundle\Validator\Constraints\RouteType: ~
