+++ src/SWP/Bundle/ContentBundle/Tests/Functional/Resources/test_file.png
